requirements.txt
python-telegram-bot==20.0
apscheduler
peewee
python-dotenv
